+++ CHANGELOG.md
@@ -6,5 +6,6 @@ All notable changes to this project will be documented in this file. This change
 - Protocol for unmanaged resources, defining a method to free the resource
 - Initialize GLFW
 - Create windows
+- Set callbacks on most types of input
 
 [Unreleased]: https://github.com/IGJoshua/s-expresso/compare/master..develop